Bug fixes
Convert unicode non-breaking spaces to ASCII spaces on handling a previewed or posted message - without this, the parser doesn't recognise smileys correctly.
Redirect drag / drop events to the plupload element if it exists - without this, dragging an image to the edit window embeds it as base64 text, crashing the browser.